Refactor: remame booleansolve_fxl --> boolean_solve_fxl
